+++ lib/src/commands/gg_kidney.dart
@@ -5,7 +5,7 @@
 // found in the LICENSE file in the root of this package.
 
 import 'package:args/command_runner.dart';
-import 'package:gg_kidney/src/commands/check_all.dart';
+import 'package:gg_kidney/src/commands/maintain.dart';
 import 'package:gg_kidney/src/commands/copy_file.dart';
 import 'package:gg_kidney/src/commands/delete_file.dart';
 import 'package:gg_kidney/src/commands/open_with_vscode.dart';
@@ -24,7 +24,7 @@ class GgKidney extends Command<dynamic> {
     addSubcommand(OpenWithVscode(ggLog: ggLog));
     addSubcommand(CopyFile(ggLog: ggLog));
     addSubcommand(DeleteFile(ggLog: ggLog));
-    addSubcommand(CheckAll(ggLog: ggLog));
+    addSubcommand(Maintain(ggLog: ggLog));
     addSubcommand(RunShellCommand(ggLog: ggLog));
     addSubcommand(Graph(ggLog: ggLog));
   }